SmoothCoreExecutorServiceFactory revision

+++ misty-smooth-core/src/main/java/org/misty/smooth/core/init/SmoothCoreExecutorServiceFactory.java
@@ -27,9 +27,9 @@ public class SmoothCoreExecutorServiceFactory implements Function<SmoothEnvironm
 
     @Override
     public ExecutorService apply(SmoothEnvironment smoothEnvironment) {
-        int corePoolSize = getCorePoolSize(smoothEnvironment);
-        int maximumPoolSize = getMaximumPoolSize(smoothEnvironment);
-        long keepAliveTime = getKeepAliveTime(smoothEnvironment);
+        short corePoolSize = getCorePoolSize(smoothEnvironment);
+        short maximumPoolSize = getMaximumPoolSize(smoothEnvironment);
+        short keepAliveTime = getKeepAliveTime(smoothEnvironment);
         TimeUnit unit = TimeUnit.SECONDS;
         BlockingQueue<Runnable> workQueue = getWorkQueue(smoothEnvironment);
         ThreadFactory threadFactory = getThreadFactory(smoothEnvironment);
@@ -38,7 +38,7 @@ public class SmoothCoreExecutorServiceFactory implements Function<SmoothEnvironm
         return new ThreadPoolExecutor(corePoolSize, maximumPoolSize, keepAliveTime, unit, workQueue, threadFactory, handler);
     }
 
-    public int getCorePoolSize(SmoothEnvironment smoothEnvironment) {
+    public short getCorePoolSize(SmoothEnvironment smoothEnvironment) {
         String key = ThreadPoolArgument.CoreSize.KEY;
         short preset = ThreadPoolArgument.CoreSize.PRESET;
         short min = ThreadPoolArgument.CoreSize.MIX;
@@ -46,7 +46,7 @@ public class SmoothCoreExecutorServiceFactory implements Function<SmoothEnvironm
         return smoothEnvironment.getValue(key, new EnvironmentShortValidator(key, preset, min, max));
     }
 
-    public int getMaximumPoolSize(SmoothEnvironment smoothEnvironment) {
+    public short getMaximumPoolSize(SmoothEnvironment smoothEnvironment) {
         String key = ThreadPoolArgument.MaxSize.KEY;
         short preset = ThreadPoolArgument.MaxSize.PRESET;
         short min = ThreadPoolArgument.MaxSize.MIX;
@@ -54,7 +54,7 @@ public class SmoothCoreExecutorServiceFactory implements Function<SmoothEnvironm
         return smoothEnvironment.getValue(key, new EnvironmentShortValidator(key, preset, min, max));
     }
 
-    public long getKeepAliveTime(SmoothEnvironment smoothEnvironment) {
+    public short getKeepAliveTime(SmoothEnvironment smoothEnvironment) {
         String key = ThreadPoolArgument.AliveSecond.KEY;
         short preset = ThreadPoolArgument.AliveSecond.PRESET;
         short min = ThreadPoolArgument.AliveSecond.MIX;
@@ -79,7 +79,7 @@ public class SmoothCoreExecutorServiceFactory implements Function<SmoothEnvironm
         short preset = ThreadPoolArgument.Rotation.PRESET;
         short min = ThreadPoolArgument.Rotation.MIX;
         short max = ThreadPoolArgument.Rotation.MAX;
-        int rotation = smoothEnvironment.getValue(key, new EnvironmentShortValidator(key, preset, min, max));
+        short rotation = smoothEnvironment.getValue(key, new EnvironmentShortValidator(key, preset, min, max));
 
         short first = 1;
         AtomicInteger count = new AtomicInteger(first);
@@ -101,7 +101,7 @@ public class SmoothCoreExecutorServiceFactory implements Function<SmoothEnvironm
         };
 
         return (runnable) -> {
-            String name = namePrefix + String.format("%04d", counter.getAsInt());
+            String name = namePrefix + String.format("%03d", counter.getAsInt());
             Thread thread = new Thread(runnable, name);
             thread.setContextClassLoader(this.executeClassLoader);
             return thread;
